feat: ✨Move SubTaskEntity definition into TaskEntity for better organization
