--- lib/features/task/domain/entities/sub_task_entity.freezed.dart
@@ -1,180 +1,0 @@
-// coverage:ignore-file
-// GENERATED CODE - DO NOT MODIFY BY HAND
-// ignore_for_file: type=lint
-// ignore_for_file: unused_element, deprecated_member_use, deprecated_member_use_from_same_package, use_function_type_syntax_for_parameters, unnecessary_const, avoid_init_to_null, invalid_override_different_default_values_named, prefer_expression_function_bodies, annotate_overrides, invalid_annotation_target, unnecessary_question_mark
-
-part of 'sub_task_entity.dart';
-
-// **************************************************************************
-// FreezedGenerator
-// **************************************************************************
-
-T _$identity<T>(T value) => value;
-
-final _privateConstructorUsedError = UnsupportedError(
-    'It seems like you constructed your class using `MyClass._()`. This constructor is only meant to be used by freezed and you are not supposed to need it nor use it.\nPlease check the documentation here for more information: https://github.com/rrousselGit/freezed#adding-getters-and-methods-to-our-models');
-
-/// @nodoc
-mixin _$SubTaskEntity {
-  String get id => throw _privateConstructorUsedError;
-  String get title => throw _privateConstructorUsedError;
-  bool get isCompleted => throw _privateConstructorUsedError;
-
-  /// Create a copy of SubTaskEntity
-  /// with the given fields replaced by the non-null parameter values.
-  @JsonKey(includeFromJson: false, includeToJson: false)
-  $SubTaskEntityCopyWith<SubTaskEntity> get copyWith =>
-      throw _privateConstructorUsedError;
-}
-
-/// @nodoc
-abstract class $SubTaskEntityCopyWith<$Res> {
-  factory $SubTaskEntityCopyWith(
-          SubTaskEntity value, $Res Function(SubTaskEntity) then) =
-      _$SubTaskEntityCopyWithImpl<$Res, SubTaskEntity>;
-  @useResult
-  $Res call({String id, String title, bool isCompleted});
-}
-
-/// @nodoc
-class _$SubTaskEntityCopyWithImpl<$Res, $Val extends SubTaskEntity>
-    implements $SubTaskEntityCopyWith<$Res> {
-  _$SubTaskEntityCopyWithImpl(this._value, this._then);
-
-  // ignore: unused_field
-  final $Val _value;
-  // ignore: unused_field
-  final $Res Function($Val) _then;
-
-  /// Create a copy of SubTaskEntity
-  /// with the given fields replaced by the non-null parameter values.
-  @pragma('vm:prefer-inline')
-  @override
-  $Res call({
-    Object? id = null,
-    Object? title = null,
-    Object? isCompleted = null,
-  }) {
-    return _then(_value.copyWith(
-      id: null == id
-          ? _value.id
-          : id // ignore: cast_nullable_to_non_nullable
-              as String,
-      title: null == title
-          ? _value.title
-          : title // ignore: cast_nullable_to_non_nullable
-              as String,
-      isCompleted: null == isCompleted
-          ? _value.isCompleted
-          : isCompleted // ignore: cast_nullable_to_non_nullable
-              as bool,
-    ) as $Val);
-  }
-}
-
-/// @nodoc
-abstract class _$$SubTaskEntityImplCopyWith<$Res>
-    implements $SubTaskEntityCopyWith<$Res> {
-  factory _$$SubTaskEntityImplCopyWith(
-          _$SubTaskEntityImpl value, $Res Function(_$SubTaskEntityImpl) then) =
-      __$$SubTaskEntityImplCopyWithImpl<$Res>;
-  @override
-  @useResult
-  $Res call({String id, String title, bool isCompleted});
-}
-
-/// @nodoc
-class __$$SubTaskEntityImplCopyWithImpl<$Res>
-    extends _$SubTaskEntityCopyWithImpl<$Res, _$SubTaskEntityImpl>
-    implements _$$SubTaskEntityImplCopyWith<$Res> {
-  __$$SubTaskEntityImplCopyWithImpl(
-      _$SubTaskEntityImpl _value, $Res Function(_$SubTaskEntityImpl) _then)
-      : super(_value, _then);
-
-  /// Create a copy of SubTaskEntity
-  /// with the given fields replaced by the non-null parameter values.
-  @pragma('vm:prefer-inline')
-  @override
-  $Res call({
-    Object? id = null,
-    Object? title = null,
-    Object? isCompleted = null,
-  }) {
-    return _then(_$SubTaskEntityImpl(
-      id: null == id
-          ? _value.id
-          : id // ignore: cast_nullable_to_non_nullable
-              as String,
-      title: null == title
-          ? _value.title
-          : title // ignore: cast_nullable_to_non_nullable
-              as String,
-      isCompleted: null == isCompleted
-          ? _value.isCompleted
-          : isCompleted // ignore: cast_nullable_to_non_nullable
-              as bool,
-    ));
-  }
-}
-
-/// @nodoc
-
-class _$SubTaskEntityImpl implements _SubTaskEntity {
-  const _$SubTaskEntityImpl(
-      {required this.id, required this.title, required this.isCompleted});
-
-  @override
-  final String id;
-  @override
-  final String title;
-  @override
-  final bool isCompleted;
-
-  @override
-  String toString() {
-    return 'SubTaskEntity(id: $id, title: $title, isCompleted: $isCompleted)';
-  }
-
-  @override
-  bool operator ==(Object other) {
-    return identical(this, other) ||
-        (other.runtimeType == runtimeType &&
-            other is _$SubTaskEntityImpl &&
-            (identical(other.id, id) || other.id == id) &&
-            (identical(other.title, title) || other.title == title) &&
-            (identical(other.isCompleted, isCompleted) ||
-                other.isCompleted == isCompleted));
-  }
-
-  @override
-  int get hashCode => Object.hash(runtimeType, id, title, isCompleted);
-
-  /// Create a copy of SubTaskEntity
-  /// with the given fields replaced by the non-null parameter values.
-  @JsonKey(includeFromJson: false, includeToJson: false)
-  @override
-  @pragma('vm:prefer-inline')
-  _$$SubTaskEntityImplCopyWith<_$SubTaskEntityImpl> get copyWith =>
-      __$$SubTaskEntityImplCopyWithImpl<_$SubTaskEntityImpl>(this, _$identity);
-}
-
-abstract class _SubTaskEntity implements SubTaskEntity {
-  const factory _SubTaskEntity(
-      {required final String id,
-      required final String title,
-      required final bool isCompleted}) = _$SubTaskEntityImpl;
-
-  @override
-  String get id;
-  @override
-  String get title;
-  @override
-  bool get isCompleted;
-
-  /// Create a copy of SubTaskEntity
-  /// with the given fields replaced by the non-null parameter values.
-  @override
-  @JsonKey(includeFromJson: false, includeToJson: false)
-  _$$SubTaskEntityImplCopyWith<_$SubTaskEntityImpl> get copyWith =>
-      throw _privateConstructorUsedError;
-}
